% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output.R
\name{plot.nbb.estimateSTS.output}
\alias{plot.nbb.estimateSTS.output}
\title{Plot function for objects of class 'nbb.estimateSTS.output'}
\usage{
\method{plot}{nbb.estimateSTS.output}(x, series_name = "", ...)
}
\arguments{
\item{x}{an object of class 'nbb.estimateSTS.output'}

\item{series_name}{a character object with the name of the series}

\item{\dots}{further arguments passed to or from other methods.}
}
\description{
Plot function for objects of class 'nbb.estimateSTS.output'
}
