% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output.R
\name{print.nbb.estimateSTS.output}
\alias{print.nbb.estimateSTS.output}
\title{Print function for objects of class 'nbb.estimateSTS.output'}
\usage{
\method{print}{nbb.estimateSTS.output}(x, ...)
}
\arguments{
\item{x}{an object of class 'nbb.estimateSTS.output'}

\item{\dots}{further arguments passed to or from other methods.}
}
\description{
Print function for objects of class 'nbb.estimateSTS.output'
}
