% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output.R
\name{summary.nbb.estimateSTS.output}
\alias{summary.nbb.estimateSTS.output}
\title{Summary function for objects of class 'nbb.estimateSTS.output'}
\usage{
\method{summary}{nbb.estimateSTS.output}(x, ...)
}
\arguments{
\item{x}{an object of class 'nbb.estimateSTS.output'}

\item{\dots}{further arguments passed to or from other methods.}
}
\description{
Summary function for objects of class 'nbb.estimateSTS.output'
}
